+++ openmls/Cargo.toml
@@ -35,7 +35,6 @@ criterion = "^0.3"
 flexi_logger = "0.18"
 itertools = "0.10"
 test_macros = { path = "../test_macros" }
-test_utils = { path = "../test_utils" }
 openmls = { path = ".", features = ["expose-test-vectors"] }
 
 [[bench]]
