+++ openmls/src/test_utils/mod.rs
@@ -7,6 +7,8 @@ use std::{
     io::{BufReader, Write},
 };
 
+pub mod test_framework;
+
 pub fn write(file_name: &str, obj: impl Serialize) {
     let mut file = match File::create(file_name) {
         Ok(f) => f,
